requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
appnope==0.1.4
asttokens==2.4.1
audioread==3.0.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.1
contourpy==1.2.0
cramjam==2.8.2
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
fastparquet==2024.2.0
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
grpcio==1.62.2
idna==3.6
ipykernel==6.29.3
ipython==8.22.2
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mido==1.3.2
mpmath==1.3.0
msgpack==1.0.8
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
omegaconf==2.3.0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
pretty-midi==0.2.10
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.1
pycparser==2.21
Pygments==2.17.2
pynput==1.7.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-rtmidi==1.5.8
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==69.1.1
simpleaudio @ git+https://github.com/cexen/py-simple-audio.git@6a7cb95c5af4537bad72bad9b190e09cb6d7883c
six==1.16.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
stack-data==0.6.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.4.0
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.2
widgetsnbextension==4.0.10
zipfile36==0.1.3
